Coordinates output in .csv file and intermediate result plots changed.
Updated getTransform.m to output a .csv with the reference, warped, and unwarped coordinates, as well as the coordinates of a warped and unwarped meshgrid to show the result of the transform as a vector field.
The script now also only displays intermediate result plots if a user-defined flag is set.
The plots of the registration middle steps and the output can now be done via an R script.